+++ Cesium3DTilesSelection/src/LayerJsonTerrainLoader.cpp
@@ -1141,6 +1141,8 @@ CesiumAsync::Future<TileLoadResult> LayerJsonTerrainLoader::upsampleParentTile(
         auto model = upsampleGltfForRasterOverlays(
             parentModel,
             tileID,
+            false,
+            "_CESIUMOVERLAY_",
             textureCoordinateIndex);
         if (!model) {
           return TileLoadResult::createFailedResult(nullptr);